requirements.txt
Django==1.8.3
mysql-python==1.2.5
djangorestframework==3.1.3